requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.6
Flask-Login==0.6.2
Jinja2==3.1.2
Werkzeug==2.3.6
requests==2.28.0
PyYAML==5.3.1
cryptography==3.4.7
